Your commit message here

+++ backend/routers/posts.py
@@ -10,6 +10,10 @@ from database import get_db,Base, engine
 from schemas import PostCreate, PostResponse,PostUpdate
 #  basiclyy pydantic define out data conaratc define what data comes in what data goes out
 #  use this for data validation, serilization and documentation for example
+
+from auth import CurrentUser
+# get cutrrent user dependecy 
+
 router = APIRouter()
 
 
@@ -33,20 +37,23 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     status_code=status.HTTP_201_CREATED,
     # Restfull best pratice when our post is created i.e. new resource are sucesfull created
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User).where(models.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(post: PostCreate, current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
+    # result = await db.execute(
+    #     select(models.User).where(models.User.id == post.user_id),
+    # )
+    # user = result.scalars().first()
+    # if not user:
+    #     raise HTTPException(
+    #         status_code=status.HTTP_404_NOT_FOUND,
+    #         detail="User not found",
+    #     )
+    # DELETED NOT NEED we have add current_user dependecny 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        # user_id=post.user_id,
+        user_id=current_user.id,
+
     )
     db.add(new_post)
     await db.commit()
@@ -85,26 +92,33 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 async def update_post_full(
     post_id: int,
     post_data: PostCreate,
+    current_user:CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
-    if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
-        # check user exists 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-                   select(models.User).where(models.User.id == post_data.user_id),
-               )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )       
+    # NOT NEED HAVE ADDED current_user Dependecy 
+    # if not post:
+    #     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    #     # check user exists 
+    # if post_data.user_id != post.user_id:
+    #     result = await db.execute(
+    #                select(models.User).where(models.User.id == post_data.user_id),
+    #            )
+    #     user = result.scalars().first()
+    #     if not user:
+    #         raise HTTPException(
+    #             status_code=status.HTTP_404_NOT_FOUND,
+    #             detail="User not found",
+    #         )  
+    #  check code belong to current user
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )    
     post.title=post_data.title
     post.content=post_data.content
-    post.user_id=post_data.user_id
 
 
     await db.commit()
@@ -120,6 +134,7 @@ async def update_post_full(
 async def update_post_partial(
     post_id: int,
     post_data: PostUpdate,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -127,6 +142,15 @@ async def update_post_partial(
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
         # check user exists 
+
+    # user check
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )  
+
+
     update_data=post_data.model_dump(exclude_unset=True) # this will give dictonary 
 #  IF with out exclude_unset if person do title update pydentic will send all other data alogn with that 
 # with exclude_unset true we only send what klient want to udate 
@@ -143,14 +167,18 @@ async def update_post_partial(
 
 # DELETE POST
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
         # check user exists 
-
-
+    # post ownership check   
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )  
 
         
     await db.delete(post)
